klp_test_support_mod: add interface for adding {k,kret,j}probes
In order to facilitate testing for bad interactions between live patching
and kprobes, add a debugfs interface for adding the latter.

New debugfs files are

  klp_test_support/add_kprobe
  klp_test_support/add_kretprobe
  klp_test_support/add_jprobe
  klp_test_support/remove_probes
  klp_test_support/active_probes

Writing a function name to any of the first three adds a probe of
the respective kind. Writing a function name to 'remove_probes' uninstall
all probes on that function. Finally, 'active_probes' shows a list of
probes created through this interface.

Signed-off-by: Nicolai Stange <nstange@suse.de>
Signed-off-by: Libor Pechacek <lpechacek@suse.com>