Generate linux version code in a portable way
Commit 28279223239e ("Fix 'mapfile: -d: invalid option' error") introduced
breakage which is not detected by Bash errexit behavior. The symptom is the
following line in script output:

./klp_tc_functions.sh: line 320: 5 4 0: syntax error in expression (error token is "4 0")

The root cause is that single space does not serve as a field delimiter for
mapfile. At this point I stop experimenting with this Bash specific command and
adopt portable version suggested by Petr Vorel.

Fixes: 2827922 ("Fix 'mapfile: -d: invalid option' error")
Suggested-by: Petr Vorel <pvorel@suse.cz>
Signed-off-by: Libor Pecháček <lpechacek@suse.cz>