klp_tc_17: Avoid running the test on s390x
Test case 17 checks that patching a kprobed function fails. Both kprobe and
livepatch set IPMODIFY ftrace ops flag which is exclusive, so the two cannot
coexist on one function. However, the architecture needs to have
HAVE_KPROBES_ON_FTRACE for kprobe to use ftrace. s390 does not have it
(contrary to x86 and powerpc). The test case thus fails, but there is no issue
with the implementation. Quite the opposite.

Closes GH-7

Reported-by: Miroslav Beneš <mbenes@suse.cz>
Signed-off-by: Libor Pecháček <lpechacek@suse.cz>
Reviewed-by: Petr Vorel <pvorel@suse.cz>