Wait for module reference count to drain after patch deactication
After a livepatch transition to the disabled state has finished, the
reference on the module will get released asynchronously from a scheduled
work. In fact rmmod has been observed to occasionally report EBUSY even
though klp_wait_complete() completed successfully.

In order to avoid spuriously failing tests, make klp_wait_complete() to
also wait for the module refcount to drop to zero after an unpatching
transition has completed. To this end, ass the livepatch module name of
interest as an additional argument to klp_wait_complete(). In addition
to waiting for the transition to complete (as before), let it also wait for
the module reference count to drain to zero in case the livepatch is found
to have been disabled.

Signed-off-by: Nicolai Stange <nstange@suse.de>