klp_tc_8: Wait for module reference count to drain after replace_all patch is applied
Commit 619c4855a6dd ("Wait for module reference count to drain after patch
deactication") fixed a race condition when a live patch was disabled and the
testing framework did not wait for its refcount to drop so it led to test failures.

There is still a similar problem in klp_tc_8.sh test for replace_all live
patches. When a replace_all patch is applied, module_put() is eventually called
for all present live patches. The test does not wait for it and might, once
again, lead to a failure.

Call klp_wait_complete() even for removing all getpid patches which were
replaced.

Signed-off-by: Miroslav Benes <mbenes@suse.cz>